refactor: split git metrics into per-metric files
Each git metric provider now lives in its own file for easier
navigation. Shared infrastructure (constants, IsGitMetric,
loadGitMetric, loadGitMeasureMetric, walkGitFiles) remains
in metrics.go.

Fixes #134
